Move *Envelope to geometryObjects folder and update all dependencies. Add license header for spatialPartitioning/Octree*.scala.
