+++ src/main/scala/com/spark3d/geometryObjects/ShellEnvelope.scala
@@ -13,10 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.spark3d.geometry
+package com.spark3d.geometryObjects
 
 import scala.math._
-import com.spark3d.geometryObjects._
 
 /** Defines a shell of 3D coordinate space. Shell here is made by a difference of
   * of two two concentric spheres. This can be used to define a bounding box of a geometryObject
