+++ src/test/scala/com/spark3d/geometryObjects/Point3DTest.scala
@@ -15,9 +15,7 @@
  */
 package com.spark3d.geometryObjects
 
-import com.spark3d.geometry.BoxEnvelope
 import org.scalatest.{BeforeAndAfterAll, FunSuite}
-import com.spark3d.geometryObjects._
 import com.spark3d.geometryObjects.Shape3D._
 
 /**
